refactor: changed the name of main.tsx to IndividualRoleListing.tsx
