+++ frontend/src/routes.tsx
@@ -1,6 +1,6 @@
 import type { RouteObject } from "react-router"
 import AllRoleListing from "./pages/applicant/AllRoleListing"
-import ViewRoleListing from "./pages/applicant/main"
+import ViewRoleListing from "./pages/applicant/IndividualRoleListing"
 import CreateRoleListing from "./pages/hr/CreateRoleListing"
 
 const routes: RouteObject[] = [
